integrated everything into one file for now
